Top 10 causes of death in the accumulated data
We obtained the ICD-10 of the first 10 keys of the death_cause_order dictionary, which was used to find the top 10 causes of death, which were saved in a list. After that, a second list containing positions 1 to 10 was used, and with a function a list of tuples was generated, which was printed in a pandas dataframe.